added authentication to deleting a user and user messages

+++ app.py
@@ -264,16 +264,35 @@ def delete_user():
 
     form = LoginForm()
 
+    #FIXME: why is our flash showing up on homepage after deleting an account
+
     flash(
-        "Danger, are you sure you want to delete your account? Login if so," +
-        "click cancel if not. All user generated messages will also be deleted.",
+        "This action cannot be undone. Are you sure you would like to" +
+        "delete your account and messages?",
         'danger'
     )
     if form.validate_on_submit():
         # TODO: Validate login and Delete all user messages
-        do_logout()
-        db.session.delete(g.user)
-        db.session.commit()
+        user = User.authenticate(
+            form.username.data,
+            form.password.data,
+        )
+
+        if user:
+            user_messages = g.user.messages
+            for message in user_messages:
+                db.session.delete(message)
+                db.session.commit()
+            do_logout()
+            db.session.delete(g.user)
+            db.session.commit()
+            return redirect('/')
+
+        else:
+            flash("Invalid credentials.", 'danger')
+
+
+
 
     return render_template("users/login-to-delete.html", form=form)
